refactor: utils 폴더 shared 하위로 이동
