+++ docs/checklist.md
@@ -78,7 +78,7 @@
 
 ### 로컬스토리지 서비스
 
-- [ ] 데이터 타입 정의
+- [x] 데이터 타입 정의
 - [x] Create 함수
 - [x] Read 함수
 - [ ] Update 함수
