+++ frontend/src/shared/utils/console.ts
