+++ frontend/src/shared/utils/format.ts
