+++ frontend/src/shared/utils/uuid.ts
